webapp/requirements/base.txt
django==3.2.5
psycopg2==2.9.1
Pillow==8.3.1
djangorestframework==3.12.4
django-rest-swagger
drf-yasg